# [Enhancement] Optimize the performance of calc mem usage for bitmap column