Why I'm doing:

BitmapValue's calculation of mem usage will consume large CPU, and every time accumulator appends a chunk, it needs to calculate all the value of bitmap column.

What I'm doing:

Incrementally calculate mem usage of chunk

Benchmark case: https://github.com/StarRocks/StarRocksBenchmark/pull/471

Before Optimization: 73s

* CHUNK_ACCUMULATE: 59.5s
* OtherJoinConjunctEvaluateTime: 8s372ms
* OutputBuildColumnTime: 3s42ms
* SearchHashTableTime: 1s377ms

After Optimization: 21.8s

* CHUNK_ACCUMULATE: 8.1s
* OtherJoinConjunctEvaluateTime: 8s361ms
* OutputBuildColumnTime: 3s65ms
* SearchHashTableTime: 1s343ms

```
mysql> select count(*) from lineorder;                                                                                                                                                                                                                                            
+-----------+                                                                                                                                                                                                                                                                     
| count(*)  |                                                                                                                                                                                                                                                                     
+-----------+                                                                                                                                                                                                                                                                     
| 143999468 |                                                                                                                                                                                                                                                                     
+-----------+

mysql> select distinct(lo_linenumber) from lineorder;                                                                                     
+---------------+
| lo_linenumber |
+---------------+
|             4 |
|             1 |
|             5 |
|             2 |
|             6 |
|             3 |
|             7 |
+---------------+

mysql> select k1, bitmap_count(v1) from bitmap_1m where k1>=1 and k1<=7;
+------+------------------+
| k1   | bitmap_count(v1) |
+------+------------------+
|    2 |           239065 |
|    1 |           239149 |
|    3 |           239137 |
|    4 |           239042 |
|    5 |           238654 |
|    6 |           239415 |
|    7 |           239186 |
+------+------------------+

select count(lo_partkey), count(v1) from lineorder join bitmap_1m on lo_linenumber=k1 and bitmap_contains(v1, lo_orderkey);
```

## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with different default values
- [ ] Policy changes: use new policy to replace old one, functionality automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [x] I have checked the version labels which the pr will be auto-backported to the target branch
  - [x] 3.2
  - [x] 3.1
  - [x] 3.0
  - [x] 2.5